Separated LaunchFileReader and ROS1LaunchFileReader
